+++ Cargo.toml
@@ -10,7 +10,9 @@ edition = "2018"
 
 [target.'cfg(windows)'.dependencies]
 winapi = { version = "0.3", features = ["powerbase", "winnt", "std", "winbase"] }
-iced = "0.1"
+iced = "0.2"
+log = "0.4"
+env_logger = "0.8.3"
 
 [[bin]]
 name = "nosleep"
