+++ src/lib.rs
@@ -3,10 +3,10 @@ mod windows;
 
 #[cfg(target_os = "windows")]
 pub fn prevent_sleep() {
-    windows::prevent_sleep();
+    windows::system::prevent_sleep();
 }
 
 #[cfg(target_os = "windows")]
 pub fn default_sleep() {
-    windows::default_sleep();
+    windows::system::default_sleep();
 }
