+++ src/main.rs
@@ -1,3 +1,4 @@
+#[macro_use] extern crate log;
 use ::nosleep;
 
 fn main() {
